refactor
- rename "Plane" mode to "Lines"
- add more parameters to WingsLines (lines count & angles settings)
- update preset
- add "noDepthTest" parameter to drawPlane()
